requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.43.3
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
datasets==4.0.0
diffusers @ git+https://github.com/huggingface/diffusers.git@f20aba3e87662735726e1680acd27edecef05a7f
dill==0.3.8
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.5.0
ftfy==6.3.1
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==3.1.0
packaging==25.0
pandas==2.3.1
peft==0.17.0
pillow==10.4.0
piq==0.8.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scipy==1.13.1
six==1.17.0
sympy==1.13.3
timm==1.0.19
tokenizers==0.21.4
torch==2.4.0
torch-fidelity==0.3.0
torchvision==0.19.0+cu121
tqdm==4.67.1
transformers==4.55.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
